document summarizer/requirements.txt
Flask
pdfplumber
pandas
sumy
fpdf
python-docx
